Moved code in TensorFlow.{h,cpp} to the CTensorFLow  module (ctensorflow_init.{h,cpp}), so that the stdlib/public/runtime/ need not  depend on the TF library.
Also addresed more feedback.
