--- include/swift/Runtime/TensorFlow.h
@@ -1,30 +1,0 @@
-//===--- TensorFlow.h - Runtime support for TensorFlow ----------*- c++ -*-===//
-//
-// This source file is part of the Swift.org open source project
-//
-// Copyright (c) 2014 - 2017 Apple Inc. and the Swift project authors
-// Licensed under Apache License v2.0 with Runtime Library Exception
-//
-// See https://swift.org/LICENSE.txt for license information
-// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
-//
-//===----------------------------------------------------------------------===//
-//
-// SWIFT_ENABLE_TENSORFLOW
-//
-//===----------------------------------------------------------------------===//
-
-#include "swift/Runtime/Config.h"
-#include <cstdint>
-
-namespace swift {
-
-// Caller owns the returned tensor.
-// TODO: Generalize to create tensors from other shapes and dtypes.
-SWIFT_RUNTIME_EXPORT void *swift_tfc_CreateScalarFloatTensor(int32_t val);
-
-SWIFT_RUNTIME_EXPORT
-void swift_tfc_TFE_Execute(void *op, void **retvals, int32_t *num_retvals,
-                           void *status);
-
-} // end namespace swift
