+++ stdlib/public/CTensorFlow/ctensorflow_init.h
@@ -1,6 +1,8 @@
 #ifndef SWIFT_SRC_SWIFT_STDLIB_CTENSORFLOW_INIT_H_
 #define SWIFT_SRC_SWIFT_STDLIB_CTENSORFLOW_INIT_H_
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -13,6 +15,17 @@ extern "C" {
 // also dumps verbose logs at that level.
 extern void InitTensorFlowRuntime(unsigned char enable_debug_logging,
                                   int verbose_level);
+
+//===----------------------------------------------------------------------===//
+// - MARK: Runtime functions to be called via IRGen.
+//===----------------------------------------------------------------------===//
+
+// Caller owns the returned tensor.
+// TODO: Generalize to create tensors from other shapes and dtypes.
+void *swift_tfc_CreateScalarFloatTensor(int32_t val);
+
+void swift_tfc_TFE_Execute(void *op, void **retvals, int32_t *num_retvals,
+                           void *status);
 
 #ifdef __cplusplus
 } /* end extern "C" */
